+++ packages/design-system/src/stories/Skeleton.stories.tsx
@@ -11,7 +11,7 @@ import {
 	StackVertical,
 } from '../';
 
-import SkeletonPrimitive from '..//components/Skeleton/Primitive/Skeleton.Primitive';
+import SkeletonPrimitive from '../components/Skeleton/Primitive/Skeleton.Primitive';
 
 export default {
 	title: 'Feedback/Skeleton',