+++ packages/storybook/src/docs/BadgeStorybook.tsx
@@ -1,5 +1,4 @@
 import { memo } from 'react';
-import type { FunctionComponent } from 'react';
 
 import Badge, { BadgeProps } from './Badge';