+++ packages/storybook/src/docs/FigmaIframe.tsx
@@ -10,24 +10,26 @@ const iframeProps = {
 
 const FigmaIframe = ({ light, dark, ...rest }: { light?: string; dark?: string }) => (
 	<>
-		{light && (
+		{light ? (
 			<iframe
+				title="Figma"
 				{...iframeProps}
 				{...rest}
 				className={classnames(styles.iframe, 'figma-iframe figma-iframe--light')}
 				src={`https://www.figma.com/embed?embed_host=storybook&url=\
                   ${light}`}
 			/>
-		)}
-		{dark && (
+		) : null}
+		{dark ? (
 			<iframe
+				title="Figma"
 				{...iframeProps}
 				{...rest}
 				className={classnames(styles.iframe, 'figma-iframe figma-iframe--dark')}
 				src={`https://www.figma.com/embed?embed_host=storybook&url=\
 				  ${dark}`}
 			/>
-		)}
+		) : null}
 	</>
 );